Add trips

+++ Navtrack.Api.Model/Assets/GetTripsModel.cs
@@ -5,7 +5,7 @@ namespace Navtrack.Api.Model.Assets
 {
     public class GetTripsModel : ResultsModel<TripModel>
     {
-        public double TotalDistance => Results.Sum(x => x.Distance);
+        public int TotalDistance => Results.Sum(x => x.Distance);
         public int TotalLocations => Results.Sum(x => x.Locations.Count);
     }
 }
